+++ src/command_modules/azure-cli-vm/azure/cli/command_modules/vm/_template_builder.py
@@ -4,86 +4,11 @@
 # --------------------------------------------------------------------------------------------
 
 
-from collections import OrderedDict
-import json
-
 from enum import Enum
 
 from azure.cli.core.util import b64encode
 from azure.cli.core.profiles import ResourceType
-
-
-class ArmTemplateBuilder(object):
-
-    def __init__(self):
-        template = OrderedDict()
-        template['$schema'] = \
-            'https://schema.management.azure.com/schemas/2015-01-01/deploymentTemplate.json#'
-        template['contentVersion'] = '1.0.0.0'
-        template['parameters'] = {}
-        template['variables'] = {}
-        template['resources'] = []
-        template['outputs'] = {}
-        self.template = template
-
-        parameters = OrderedDict()
-        parameters['$schema'] = \
-            'https://schema.management.azure.com/schemas/2015-01-01/deploymentParameters.json#'
-        parameters['contentVersion'] = '1.0.0.0'
-        parameters['parameters'] = {}
-        self.parameters = parameters
-
-    def add_resource(self, resource):
-        self.template['resources'].append(resource)
-
-    def add_variable(self, key, value):
-        self.template['variables'][key] = value
-
-    def add_parameter(self, key, value):
-        self.template['parameters'][key] = value
-
-    def add_secure_parameter(self, key, value, description=None):
-        param = {
-            "type": "securestring",
-            "metadata": {
-                "description": description or 'Secure {}'.format(key)
-            }
-        }
-        self.template['parameters'][key] = param
-        self.parameters['parameters'][key] = {'value': value}
-
-    def add_id_output(self, key, provider, property_type, property_name):
-        new_output = {
-            key: {
-                'type': 'string',
-                'value': "[resourceId('{}/{}', '{}')]".format(
-                    provider, property_type, property_name)
-            }
-        }
-        self.template['outputs'].update(new_output)
-
-    def add_output(self, key, property_name, provider=None, property_type=None,
-                   output_type='string', path=None):
-
-        if provider and property_type:
-            value = "[reference(resourceId('{provider}/{type}', '{property}'),providers('{provider}', '{type}').apiVersions[0])".format(  # pylint: disable=line-too-long
-                provider=provider, type=property_type, property=property_name)
-        else:
-            value = "[reference('{}')".format(property_name)
-        value = '{}.{}]'.format(value, path) if path else '{}]'.format(value)
-        new_output = {
-            key: {
-                'type': output_type,
-                'value': value
-            }
-        }
-        self.template['outputs'].update(new_output)
-
-    def build(self):
-        return json.loads(json.dumps(self.template))
-
-    def build_parameters(self):
-        return json.loads(json.dumps(self.parameters))
+from azure.cli.core.commands.arm import ArmTemplateBuilder
 
 
 # pylint: disable=too-few-public-methods
@@ -815,7 +740,7 @@ def build_vmss_resource(cmd, name, naming_prefix, location, tags, overprovision,
         'computerNamePrefix': naming_prefix,
         'adminUsername': admin_username
     }
-    if authentication_type == 'password':
+    if authentication_type == 'password' and admin_password:
         os_profile['adminPassword'] = "[parameters('adminPassword')]"
     else:
         os_profile['linuxConfiguration'] = {