+++ src/command_modules/azure-cli-vm/HISTORY.rst
@@ -4,6 +4,7 @@ Release History
 ===============
 2.0.25
 ++++++
+* vm/vmss create: ensure commands can run under proxy with unsigned certificates.
 * vmss:(PREVIEW) support low priority
 
 2.0.24