+++ src/command_modules/azure-cli-vm/HISTORY.rst
@@ -7,6 +7,10 @@ Release History
 * vm/vmss create: ensure commands can run under proxy with unsigned certificates.
 * vmss:(PREVIEW) support low priority
 
+2.0.25
+++++++
+* `vm/vmss create` - `--admin-password` updated to type secureString.
+
 2.0.24
 ++++++
 * vmss:(PREVIEW) cross zone support